refactor: extract useClaimReward
