+++ src/components/UserStatus/RewardCard.tsx
@@ -1,50 +1,19 @@
 import { Box, Button, HStack, Icon, Stack, Text } from '@chakra-ui/react'
 import { t } from '@lingui/macro'
-import { produce } from 'immer'
-import { memo, useMemo, useState } from 'react'
-import { TransactionExecutionError, UserRejectedRequestError, formatUnits } from 'viem'
-import {
-  useAccount,
-  useBalance,
-  useChainId,
-  useConfig,
-  useReadContract,
-  useSwitchChain,
-  useToken,
-  useWriteContract,
-} from 'wagmi'
-import { waitForTransactionReceipt } from 'wagmi/actions'
+import { memo, useMemo } from 'react'
+import { formatUnits } from 'viem'
+import { useAccount, useBalance, useReadContract, useToken } from 'wagmi'
 import { rewardABI } from '../../abis/reward'
 import QuestionSVG from '../../assets/question.svg?react'
-import { queryClient } from '../../configs/queryClient.ts'
 import { formatMarketCap } from '../../helpers/formatMarketCap.ts'
-import { formatNumber } from '../../helpers/formatNumber'
-import { resolveTxLink } from '../../helpers/resolveTxLink.ts'
-import { useHandleError } from '../../hooks/useHandleError.ts'
-import { useToast } from '../../hooks/useToast'
-import { useUserInfo } from '../../hooks/useUserInfo'
-import { resultModal } from '../../modals/ResultModal'
 import { usePoolStore } from '../../store/poolStore'
 import { UserInfo } from '../../types/api'
+import { ProgressiveText } from '../ProgressiveText.tsx'
 import { Spinner } from '../Spinner'
 import { TokenIcon } from '../TokenIcon'
 import { Tooltip } from '../Tooltip'
-import { TxToastDescription } from '../TxToastDescription.tsx'
 import { ActionCard, ActionCardProps } from './ActionCard'
-import { ProgressiveText } from '../ProgressiveText.tsx'
-
-function clearReward(address: string, rewardPoolId: number) {
-  queryClient.setQueriesData<UserInfo>({ queryKey: ['user-info', address] }, (info) => {
-    if (!info) return info
-    return produce(info, (draft) => {
-      const rewardPool = draft.reward_pool.find((x) => x.reward_pool_id === rewardPoolId)
-      if (rewardPool) {
-        rewardPool.amount = '0'
-        rewardPool.big_amount = '0'
-      }
-    })
-  })
-}
+import { useClaimReward } from './useClaimReward.tsx'
 
 interface Props extends ActionCardProps {
   tokenIcon?: string
@@ -62,22 +31,18 @@ export const RewardCard = memo(function RewardCard({
 }: Props) {
   reward?.reward_pool_id
   const account = useAccount()
-  const currentChainId = useChainId()
-  const config = useConfig()
-  const { chainId, poolId, rewardAddress } = usePoolStore()
-  const [waiting, setWaiting] = useState(false)
-  const { switchChainAsync, isPending: switchingChain } = useSwitchChain()
-  const { writeContractAsync, isPending } = useWriteContract()
-  const { data: userInfo, isLoading: loadingUserInfo } = useUserInfo()
-  const toast = useToast({ title: t`Claim` })
+  const { chainId, rewardAddress } = usePoolStore()
   const { data: token } = useToken({ chainId, address: reward?.address })
-  const handleError = useHandleError()
   const { data: userReward, isLoading: loadingRewards } = useReadContract({
     chainId,
     abi: rewardABI,
     address: rewardAddress,
     functionName: 'userRewards',
     args: reward?.reward_pool_id && account.address ? [reward.reward_pool_id, account.address] : undefined,
+  })
+
+  const [{ switchingChain, claiming, loadingUserInfo, waiting }, claimReward] = useClaimReward(reward, {
+    defaultSymbol: defaultSymbol || token?.symbol,
   })
 
   const { data, isLoading: loadingRewardBalance } = useBalance({ chainId, address: reward?.address })
@@ -122,90 +87,18 @@ export const RewardCard = memo(function RewardCard({
           alignSelf="stretch"
           color="neutrals.9"
           rightIcon={
-            unlocked || isPending ? undefined : (
+            unlocked || claiming ? undefined : (
               <Tooltip label={t`You can claim after the event ends.`} shouldWrapChildren>
                 <Icon as={QuestionSVG} w="initial" h="initial" />
               </Tooltip>
             )
           }
           className="purple-gradient-button"
-          onClick={async () => {
-            if (!reward || !userInfo || !rewardAddress) return
-            const rewardPool = userInfo.reward_pool.find((x) => x.reward_pool_id === reward?.reward_pool_id)
-            if (!rewardPool) {
-              toast({
-                status: 'error',
-                title: t`No matched pool found.`,
-              })
-              return
-            }
-            try {
-              if (currentChainId !== chainId) {
-                await switchChainAsync({ chainId })
-              }
-              toast({
-                status: 'loading',
-                description: t`Confirm this transaction in your wallet.`,
-              })
-              const hash = await writeContractAsync({
-                abi: rewardABI,
-                address: rewardAddress,
-                functionName: 'claim',
-                args: [reward.reward_pool_id, BigInt(reward.big_amount), rewardPool.proof],
-              })
-              const txLink = resolveTxLink(chainId, hash)
-              toast({
-                status: 'loading',
-                description: <TxToastDescription link={txLink} text={t`Transaction submitted!`} />,
-              })
-
-              setWaiting(true)
-              const receipt = await waitForTransactionReceipt(config, {
-                hash,
-                confirmations: 1,
-              })
-              setWaiting(false)
-
-              if (receipt.status === 'reverted') {
-                toast({
-                  status: 'error',
-                  description: <TxToastDescription link={txLink} text={t`Transaction failed.`} />,
-                })
-                throw new Error('The approval transaction gets reverted!')
-              } else {
-                toast({
-                  status: 'success',
-                  description: (
-                    <TxToastDescription
-                      link={txLink}
-                      text={t`Successfully claimed ${tokenSymbol?.toUpperCase()} Tokens.`}
-                      color="primary.4"
-                    />
-                  ),
-                })
-                if (account.address && poolId) clearReward(account.address, poolId)
-                await resultModal.show({
-                  title: t`Claim`,
-                  message: t`Claim Successfully`,
-                  description: t`You have successfully claimed ${formatNumber(+reward.amount, 4)} ${tokenSymbol}.`,
-                })
-              }
-            } catch (err) {
-              const cause = err instanceof TransactionExecutionError ? err.cause : err
-              if (cause instanceof UserRejectedRequestError) {
-                toast({
-                  status: 'error',
-                  description: t`Your wallet cancelled the transaction.`,
-                })
-                return
-              } else if (handleError(err)) return
-              throw err
-            } finally {
-              setWaiting(false)
-            }
+          onClick={() => {
+            claimReward()
           }}
         >
-          {isPending ? <Spinner h="24px" w="24px" color="neutrals.9" /> : t`Claim`}
+          {claiming ? <Spinner h="24px" w="24px" color="neutrals.9" /> : t`Claim`}
         </Button>
       </Stack>
     </ActionCard>
